ccl::simple-vector-delete loses with displaced arrays with a displaced-index-offset > 0.
the second loop (looping through the scratch bit vector) to copy
the items that survive into the new array misuses the offset with
the bit vector.  this causes an array out of bounds error.  instead
it should use the offset with the original array to reference the
item.

the third loop which copies the remaining items that were explicitly
excluded from the test also needs the offset to copy the right items.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@908 8bdb8f1d-59ce-db11-9e41-0016172a54ae