Revert to using "trap logical" alloc-traps on PPC32.
Rationale: the largest object we can allocate is about 2^27 bytes in size
(a DOUBLE-FLOAT vector with almost 2^24 elements).  If HEAP-IMAGE-BASE
is effectively above 2^27, we can never cause allocptr to change sign
from positive to negative (to wrap around 0), but we -might-  cause it
to change sign from negative to positive (to wrap around #x80000000).

On PPC64, we'd have to be able to guarantee that HEAP-IMAGE-BASE was
above 2^59 to ensure this.  That's hard to do, so we have to basically
limit ourselves to the low 63 bits of memory ("positive" addresses)
and do signed comparisons. (Wrapping around 0 should trap; wrapping
around #x8000000000000000 shouldn't be possible.)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@1991 8bdb8f1d-59ce-db11-9e41-0016172a54ae