In xcmain callback (both of them), get rid of code which gratuitously smacks a 0 into the top word of the value stack in the current catch frame in response to an interrupt trap (trgti nargs, 0).  (I -think- that this had something to do with reenabling interrupts in a long-forgotten cooperative scheduler.)
It's possible that until recent changes to register allocation, the
top word on the vstack at this point was either (a) always non-negative
and therefore immune to getting smacked around or (b) something not
very important.

On Linux32, this was currently a pointer to a saved special binding,
(happened to look negative because of where that stack was allocated)
and setting it to 0 caused a crash as the binding stack was unwound.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@2299 8bdb8f1d-59ce-db11-9e41-0016172a54ae