hash-find functions: FIRST-DELETED-INDEX is always a fixnum or NIL, never T. Return FIRST-DELETED-INDEX if a secondary probe finds a free entry and FIRST-DELETED-INDEX is non-nil.  (Always return an index.)
(In some cases, we were returning T instead of an index from a probe
functions.   Some callers errored trying to do (1+ T), some callers
had aggressive fixnum/optimize declarations that caused 1 to be added
to T and the result used to access the hash vector.

It's OK if the declarations are aggressive, unless they're wrong, of
course.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@2502 8bdb8f1d-59ce-db11-9e41-0016172a54ae