Start moving away from high-level "instruction templates" and towards more detailed "opcode templates".  The latter are intended to make it easier to generate code from (e.g.) the compiler.
(This is currently about half there.  The assembler doesn't work at
the moment, and the "opcode templates" are missing some things that
might be used in compiled/lap code and their definitions may contain
typos. It's worth doing, in that it'll allow the compiler to generate
machine code while bypassing much of the old implementation's parse-
time side effects and special-cases.)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@3106 8bdb8f1d-59ce-db11-9e41-0016172a54ae