Kinder, gentler CHECK-MISC-BOUND.  (What was I thinking ?)
More cleanup to FUNCALL and variants.

MAKE-TSP-CONS: fix braindamage, use only one register.
(Need to write down pc-luser-xp constraints for TSP allocation, but I
think that this meets the current artist's conception of them.)

%SYMBOL->SYMPTR does what it does on other platforms (e.g., map NIL
to a proxy symbol and/or insist on a real symbol.)  %SYMPTR->SYMVECTOR
maps something tagged as a symbol to the underlying uvector by
playing with tag bits, and %SYMVECTOR->SYMPTR does the inverse.

There were two definitions of %SCHARCODE in the file; the later
one (that took effect) was completely wrong.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@3836 8bdb8f1d-59ce-db11-9e41-0016172a54ae