RESOLVE-EEP uuo needs source and dest registers to be recoverable.
Use (int ($ #x8x)) for SLOT-UNBOUND, since it needs slot-vector, index,
and dest reg to be recoverable.

Note that the use of this range of (int N) operands may lead to confusion
on platforms that use int #x80 for syscalls.  That (a) doesn't apply to
x8664, since SYSCALL is present and used on x8664 and (b) would only
be a problem if the dest reg was %rax, which it won't be.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@4067 8bdb8f1d-59ce-db11-9e41-0016172a54ae