Don't push the function on the "value" stack when doing a MULTIPLE-VALUE-CALL, since its presence there can interfere with multiple-value continuations. (Can it be popped before recovering multiple values ?  I didn't think of that, but I suspect that that'd be simpler than what I did; check this. Maybe there is -or once was- an issue with _SPrecovervalues bashing a register.)
Pushing the function on the temp stack while collecting multiple values
avoids the issue (but it may be far better to simply vpop it before
the call to _SPrecovervalues.)

Most of the issues related to X862-MVCALL calling itself recursively
have to do with cleaning the function off of the stack in non tail-call
cases.  Since the function's not there anymore, the rationale for
splitting this into a base and recursive case isn't very strong.

Changes in X862-MVCALL exposed bugs in X862-INVOKE-FN: we don't
want to emit or otherwise reference the mvpass label if it's null.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@4085 8bdb8f1d-59ce-db11-9e41-0016172a54ae