Let's see: if thread X holds local A and tries to obtain lock B, and thread Y holds lock B and tries to obtain lock A, what is that a classic case of ?
Try to avoid that classic deadlock case in shutdown_thread_tcr() when
HAVE_TLS is in effect.  suspend_other_threads wants to hold two locks
at the same time, and that's always a bit dangerous; it might be safer to
merge TCR_LOCK and AREA_LOCK into a single lock that must be owned in order
to manipulate either the tcr or area queues.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@4526 8bdb8f1d-59ce-db11-9e41-0016172a54ae