Some extra fields in objc-message-info/objc-method-info structures.
When calculating overall alignment of a structure, treat bitfields
as if they were :unsigned-int (since there's a "containing" unsigned
int that affects alignment.  N.B: this should be imposed for anything that
has an alignment of 1, not just a size of 1.)

When loading integer types, try to use the precomputed types rather
than consing a new one.  This (a) reduces consing a little (b) is
more likely to get the alignment correct.  For (b), we really need
to distinguish between bitfields and addressable integers.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@5959 8bdb8f1d-59ce-db11-9e41-0016172a54ae