DEFSTATIC is now the preferred name for DEFGLOBAL.
Standard CL streams now bound per-thread.

Install standard foreign types here, not every time that FOREIGN-TYPES is
loaded.

Canonicalize foreign type ordinals as soon as we can.  (This is part
of a scheme to give foreign types "ordinals' that can be used to assert
pointer types; we need some of those ordinal numbers to be pre-assigned,
since we can't reference foreign types early in the cold load.)










git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@6186 8bdb8f1d-59ce-db11-9e41-0016172a54ae