Add :clozure-common-lisp to *features*.
Allow the init-file arg to startup-ccl to be a list of filenames -
the first one that exists is used.  Make the lisp-development-system
application-init-file be the list of ccl-init and openmcl-init, but
warn if openmcl-init is the one that's actually used.

Delete cocoa-ide/OpenMCL.app to minimize confusion (this means
removing :clozure-common-lisp from *features* no longer works
for the IDE, although it still works for the command line build).


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@7487 8bdb8f1d-59ce-db11-9e41-0016172a54ae