Revise behavior of command-L shortcut.
When no listener is open, command-L will create one.

If a listener window is not key, command-L will bring the closest
listener (in window z-order) to the front and make it key.

If a listener window is already key, command-L will cycle between
other open listeners.  If there are no other open listeners, it
will just beep.



git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@7489 8bdb8f1d-59ce-db11-9e41-0016172a54ae