I'm going to go ahead and commit this, even though I'm not happy with it.
I removed an old workarond for what appears to have been a Carbon bug
of some sort, so if something blows up, I guess that's a prime candidate
for examination.

I'm almost convinced that it would be advantageous  to make nib
files for editor and listener windows, just so that it's easier to
get all the details right.



git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@7495 8bdb8f1d-59ce-db11-9e41-0016172a54ae