Got tired of restarting my lisp all the time, made various hemlock printing functions bind the right magic context variables.  (This is way ugly but temporary, as the whole threading scheme needs to be replaced by something less fragile anyway, and this can be cleaned up then).
I'm checking in Matt's patch for ticket #135, because it's in my sources,
it works, and whatever else might be wrong with it is better than the
original bug.

add find-not-attribute and reverse-find-not-attribute so I don't have
to remember which way the magic arg to find-attribute goes.

make region-to-string take an optional output-string argument, return
length as second value.

Implement hi:window-buffer, also add ordered-hemlock-windows (returns
a front-to-back list of hemlock windows).


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/ccl@7528 8bdb8f1d-59ce-db11-9e41-0016172a54ae