Add :monitor-futex-wait to *known-optional-features*.
RECURSIVE-LOCK-WHOSTATE: get the whostate field, not the value field ...

Deal with whostate a little differently in FUTEX-WAIT, LOCK-FUTEX, and 
callers.

Conditionally (#+monitor-futex-wait) count futex-wait syscalls and the
time spent in them in CCL::*TOTAL-FUTEX-WAIT-CALLS* and
CCL::*TOTAL-FUTEX-WAIT-TIMES*, respectively.  (This isn't thread-safe:
two threads trying to increment these values at the same time can cause
the results to be inconsistent.)






git-svn-id: http://svn.clozure.com/publicsvn/openmcl/branches/working-0711/ccl@7925 8bdb8f1d-59ce-db11-9e41-0016172a54ae