Remaining changes to this file from ia32 branch.
Remove a whole bunch of constants that defined cpu features and opcode
modifiers.  These were derived from the GAS sources, but the way we
define opcodes doesn't use them.

Add a new pseudo operand called :self.  It's used only in one case:
(movl ($ :self) (% reg)).  The value of :self is the absolute address
of the entry point of the function being defined.

Remove the old opcode template list that was conditioned out.

Define register entries for x8632 and alter associated functions to
consult the appropriate entries based on target architecture.

Update match-template/match-template-types to exclude instructions not
applicable to the target architecture.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@10198 8bdb8f1d-59ce-db11-9e41-0016172a54ae