Use entry->addr to hide the differences between a foreign "entry" (which may be a fixnum representing a word-aligned address or an integer or something equally platform-dependent) and an address; this may make it possible to map external entry points to the libraries that define them.
Towards that end, try to use #_dladdr on Darwin.  Give up (after
much frustration) because there's no correlation between the
"handle" returned by Darwin's #_dlopen and the library base
returned by #_dladdr.  (We might be able to find a static C++
function inside dyld to do this for us, but that's pretty ugly.)

I wouldn't be at all averse to dropping (nominal) support for
10.3.9 (and therefore the deprecated, legacy dyld stuff) if
the dlfcn stuff was actually useful.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@10281 8bdb8f1d-59ce-db11-9e41-0016172a54ae