Try to keep track of the state of the insertion point in the blink-phase slot of a text-view.
Implement a "before" method on #/drawRect: for hemlock-textstorage-textview;
in it, do way to much parsing to support primitive syntax highlighting via
temporary attributes.

Handle drawing/erasing the matching paren by invalidating its glyph
rectangle; decide whether to erase it in the drawRect: method (and
use a temporary attribute when erasing it.)  Disable blinking when
the view's deactivated.  (This seems to remove the delay before the
matching paren starts blinking; I think that I've seen a delay when
reactivating the window when blinking should take place.  Sometimes.)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11037 8bdb8f1d-59ce-db11-9e41-0016172a54ae