Get rid of X862-CR-BIT-FOR-LOGICAL-COMPARISON and use the almost-identical X862-CR-BIT-FOR-UNSIGNED-COMPARISON in X862-COMPARE-LOGICAL-REGISTERS; ensure that inversion is handled there.
This seems to fix the "wrong answer" part of ticket:373, but I think
that we've got to do something on ia32 to ensure that the extra
temporary imm register gets set restored to node-ness regardless of
whether or not we branch after the compare.  (I'm also a little
concerned that we're doing the mark-as-imm too early. e.g. if X is
some expression that involves a function call and we've done
(mark-as-imm reg-for-y ...)).



git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11286 8bdb8f1d-59ce-db11-9e41-0016172a54ae