Lose the speculative MARK-AS-NODE-PRESERVING-FLAGS.
Add a MARK-TEMP1-AS-NODE-PRESERVING-FLAGS, which doesn't clobber as many flage.

Add a TEMP-POP-TEMP1-AS-UNBOXED-WORD, which sets the direction flag to indicate the unboxed state of %temp1/%edx as it pops a word into that register.

ALIGN-LOOP-HEAD was a bad idea (adds NOPs but doesn't improve performance), but attempts to actually remove uses of it failed for some unknown reason.  Make the vinsn a NOP for now.



git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11297 8bdb8f1d-59ce-db11-9e41-0016172a54ae