In ALLOC-C-FRAME and ALLOC-VARIABLE-C-FRAME, enforce a minimum frame size of 24 bytes.  GCC on Darwin seems to do this (a non-leaf C function with no locals seems to reserve 24 bytes).  If this is part of the ABI, it's not mentioned anywhere...
This issue came up in the context of this call: (#_open pathname
#$O_RDONLY).  On entry to the C library routine, the stack contains
the return address, pathname, and flags.  The routine sets a bit
in the (nominally optional) third argument (the file mode).  This
clobbers the C frame link, since we only passed two args.

Since open(2) has a variadic prototype, I suppose that in C, this would
result in there always being a 0 on the stack in the position of
the mode argument (with two-arg open), so that would make _open safe, too.
(In the three-arg case, there'd be a 0 after the third arg, of course.)



git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11414 8bdb8f1d-59ce-db11-9e41-0016172a54ae