We really want definitions of functions that return structure types to be explicit about that; at this level, we can't just translate those functions into things that accepts a pointer argument and returns NIL, since that loses information.
There was a variable - *FFI-EXPLICIT-STRUCT-RETURN* - that controlled
whether this happened or not; when it was true, the translator printed
infomation about functions that returned structures, and when it was
false, the return type was set to :VOID and an extra pointer arg was
added.  Possibly to avoid printing that cryptic-looking info, the
variable was erroneously set false at some point in the past.

(At a higher level, functions that return structures do need to accept
a pointer to a structure as their fist argument, but how that pointer's
used depends on the ABI.)

In any case. botching this means that the interfaces that we've been
using for the last year or two (whenever this change was made) are
missing struct return info for foreign functions (though it's present
for ObjC methods), so we'll need new interfaces for most platforms.
(Darwin makes heaviest use of struct return.)

Here, don't think that making struct return explicit is optional.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11470 8bdb8f1d-59ce-db11-9e41-0016172a54ae