Additional x8632 FFI details involving structure return and callbacks.
DEFINE-CALLBACK: add hair to recognize when a hidden first argument
will need to be discarded on return.  Try to deal with interactions
with the win32 _stdcall case, too.  Pass info argument to
DEFINE-CALLBACK-FUNCTION that encodes the arg discard information.

DEFINE-CALLBACK-FUNCTION: pass said info argument to
MAKE-CALLBACK-TRAMPOLINE.

MAKE-CALLBACK-TRAMPOLINE: pack arg discarding information and the
callback index into the value that gets passed to .SPcallback via
the %eax register.

.SPcallback: decode this information, and handle discarding the
appropriate number of args on return.



git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11590 8bdb8f1d-59ce-db11-9e41-0016172a54ae