Print the message before suspending other threads on entry to the kernel debugger.
Also, check to see if fd 0 is open to /dev/null on entry to the
kernel debugger (after printing the message but before suspending
threads.) If so, exit from the kernel debugger with a non-zero
value; this should cause cases where it's entered in response to
an unhandled exception to set a bit in the TCR and should cause
the exception handler to return (repeating the trap/fault).  On
Darwin, the Mach exception handler will assert that the exception
wasn't handled, causing the exception to be passed to a task-level
handler (or signal handler); for an OSX GUI application, this can
cause Apple's crash dialog to be invoked.

We could get a vaguely similar effect on other platforms by
de-installing our signal handler (or Windows exception handler),
but it wouldn't be guaranteed that the next (now fatal) exception
would be the one that we tried to propagate in this way: some 
other thread might do something innocuous - like consing - that
raises an unrelated exception.

In the case of an OSX GUI application (that hasn't redirected
its standard fds), we may be able to generate a more meaningful
crash report (with lisp thread info, lisp-aware backtraces, etc.)
than what Apple's crash dialog provides.  Apple's mechanism does
at least make it clear that the application terminated abnormally;
writing to syslog or writing to a text file somewhere may leave
some end-users (at least) puzzled ("the application terminated
unexpectedly, and I wasn't notified ?").


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11596 8bdb8f1d-59ce-db11-9e41-0016172a54ae