FreeBSD 7.1 on i386 doesn't seem able to ensure that 'current_tcr' is aligned correctly (I tried several approaches.)  It's possible that previous versions didn't do this right either, and that changes in TLS in libraries just causes misalignment.  Whatever ... Reserve a thread-local "tcrbuf" (large enough to hold a TCR + alignment overhead) and a make current_tcr be a thread-local pointer to (or slightly into) that buffer.
On FreeBSD, i386_set_fsbase() should have the side-effect of
loading the global user fs selector into %fs.  On a 64-bit
7.1 kernel, it doesn't change %fs (or at least doesn't change
it in a way that persists across context switches.)  I don't yet
know whether or not that's a bug, but for now, try to check the
value in %fs on return from the i386_set_fsbase call.  (I submitted
a FreeBSD bug report on this (amd64/130355) ; we'll see what happens.)

Note that this only affects trying to run a 32-bit CCL on a 64-bit
7.1 kernel.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11608 8bdb8f1d-59ce-db11-9e41-0016172a54ae