When trying to deterine temporary attributes in COMPUTE-TEMPORARY-ATTRIBUTES, use the #/bounds rectangle of the enclosing scroll view's content view, not the #/frame rectangle. (Note that COMPUTE-TEMPORARY-ATTRIBUTES is currently only called on window activation and when the selection changes.  That's better than setting temporary attributes in the middle of drawRect: (and causing excessive redisplay), but it means that we get no highlighting when scrolling around the buffer.
I think that what we may want to do is:

- update paren highlighting (only) when the selection changes.

- parse the whole buffer (Hemlock-wise) when it's first opened.
Add temporary attributes (syntax highlighting) for "interesting"
ranges - currently strings and ;-comments, other things 

- whenever operations that modify the buffer complete (e.g., at
#/endEditing time) re-parses the affected range (and, in some
cases, preceding and succeeding lines.)  Recompute temporary
attributes for what got re-parsed.

This would also mean that the buffer's always in a fully-parsed
state when a command runs, and that we're not affected by legacy
Hemlock limitations that try to guess how many surrounding toplevel
forms should be parsed.

I don't know if there are problems with this idea that make
it impractical, but I think that it makes more extent to explore
it than it does to try to (e.g.) intercept bounds-rectangle changes
and recompute temporary attributes while scrolling.  (There might,
of course, be other approaches.)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11616 8bdb8f1d-59ce-db11-9e41-0016172a54ae