Fix (from Shannon Spires, over a year ago) to treat Caps Lock like shift when mapping key events.  (See ticket:219.)
Subclass NSClipView; use an instance of that clip view when embedding
a hemlock text view in a scroll view.  Define a scrollToPoint: method
on that NSClipView subclass, so that we can tell when scrolling occurs
and update transient text attributes when it does.  (May need to
handle cases when the view is enlarged vertically, as well.)

Try to finally fix ticket:150 (confused by double-click after close
paren at end of file.)  There may have been similar problems having
to do with double-clicking before an open paren at the beginning
of the file; this would be caused by Hemlock's VALID-SPOT-P believing
that position 0 (at least) is in the attribute line.   (Should probably
change the Hemlock code; files don't necessarily have attribute lines,
and there are probably better ways of telling whether or not we're
in te middle of one.)  At EOF, Cocoa claims that a double-click is
trying to select a character (not a word or para); in order to
determine whether we got a single or multi-click, look at the current
event.  On Leopard at least, the proposed range is correct in this
case (and it's the granularity that's confused us); should check
Tiger.

Try to clean up some cases where paren highlighting isn't cleared
when it should be.  Wouldn't want to claim that no such cases remain,
but it's better.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11619 8bdb8f1d-59ce-db11-9e41-0016172a54ae