Write debugging/diagnostic output to the FILE* "dbgout", which is initialized to stderr.
Add a function that re-opens "dbgout" on a specified fd, add
that function to the imports table.  (This is intended to allow I/O
redirection mechanisms - like AltConsole in the IDE - to persuade
the kernel to do output to a specified FILE*/fd without having
that mechanism capture random diagnostic output to stderr, which
would still go to syslog.)

Lisp code needs to update the import table definitions and the
IDE startup code has to use the new 'open_debug_output()' mechanism.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11623 8bdb8f1d-59ce-db11-9e41-0016172a54ae