Try to work around the confusion in #/replaceCharactersInRange:withString: by adding more confusion.  This method -should- assume that the caller has done #/beginEditing on the textstorage and will do #/endEditing at some point in the future, but it calls Hemlock code that asserts that editing is not in progress.  (Some of our own methods may call this method without a preceding #/beginEditing.)  So, to kludge around this, note the edit count on the texstorage before modifying the buffer and call #/endEditing until the edit count is 0, do the Hemlock-level modification, then restore the edit count (by calling #/beginEditing to raise the edit count back to the level it had on entry.)
This is a total hack (and may break something somewhere), but it keeps
this code from generating a spurious assertion failure and allows foreign
code to modify the buffer; for a good time, try "ctrl-q option-e e" on
a US keyboard.  (This exposes other problems with non-ASCII characters
in lisp buffers; it's not clear that their syntax attributes are reasonable.)

In #/paste, assume that #/replaceCharactersInRange:withString: isn't broken,
and bracket the call to that method with #/beginEditing / #/endEditing
calls.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11903 8bdb8f1d-59ce-db11-9e41-0016172a54ae