pmcl-kernel.c: don't define __objc_personality_v0, since in the Wacky World Of Mach-O this can have the effect of shadowing the real ObjC runtime symbol.
x86-spentry64.s: in ffcall_return_registers: after saving return registers
relative to %csave0, restore them from there (not from %csave3).  Most
callers don't care (will just look at the register buffer that csave0 
pointed to), but if we're returning/raising an ObjC exception we want to get
it from the right place.

In the ObjC exception gunk, reference the new 'lisp_objc_personality' via
a 64-bit absolute address.  This seems to be OK with the Leopard linker;
we'll see about other versions.

objc-support.lisp: in SETUP-OBJC-EXCEPTION-GLOBALS, try to ensure that
this only happens once per session (since we're using a kludge to 
detect whether or not indirection is necessary.)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@11913 8bdb8f1d-59ce-db11-9e41-0016172a54ae