We can't easily support USE-VALUE, STORE-VALUE, or CONTINUE restarts after uuo-error-reg-unbound UUOs on x86(32- or 64-bit) platforms.
 - the UUO itself encodes (as #xcd #9x) the register that contains the
symbol whose value is unbound, but not the register that the (unbound)
is loaded into. (USE-VALUE and STORE-VALUE need this.)

 - the anchored UUO mechanism can make it unclear how to continue after
the UUO; we don't always follow the UUO with a meaningful instruction.

So (in this file): assume that #xcd #x90 is (the first two bytes of) a
3-byte UUO which encodes the both the source and destination register in
a third byte.  (#xcd #x90 would otherwise be a uuo-error-reg-unbound uuo
that references %eax/%rax, and I assume that we don't currently generate
such an insruction.)  Treat #xcd #x9x for non-zero x as an UNBOUND-VARIABLE,
but don't offer automatically-generated restarts.

[The plan is to transfer to the new 3-byte UUO, once all of the restart
support is in place.  Until that transfer is complete, we don't want to
offer restarts that can't work, for the reasons described above.]


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@12010 8bdb8f1d-59ce-db11-9e41-0016172a54ae