On Windows, try to obtain the UTF-16 version of argv; change some of the functions that deal with determining the image name and opening the image to accept UTF-16-encoded strings.  When the kernel's done with the image name and argv, convert the strings involved to UTF-8 (since this doesn't lose information and makes it a little easier to bootstrap changes to the lisp side of this.)
On the lisp side of this (when obtaining the heap image name and
command-line arguments), assume that the strings are UTF-8-encoded
and possibly do platform-dependent postprocessing.  (I honestly
don't know how to reliably tell what encoding was used for things
that come from the command line on other platforms; if they may
not be UTF-8, we might want to make similar changes in the kernel
to support other encodings.

This is supposed to help address ticket:475; I don't know yet if
it does, or if the kernel changes will compie on non-Windows platforms.
One way to find out ...


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@12095 8bdb8f1d-59ce-db11-9e41-0016172a54ae