#/replaceCharactersAtPosition:length:withString: UNWIND-PROTECT around #/beginEditing / #/endEditing.
Try to support :NATIVE quote-next-char behavior.  Doesn't work yet.

If our clip-view (the contentview of a scroll view) gets asked to
#/scrollToPoint: while live resize is active, don't do it.  (This
seems to prevent the unwanted scrolling that occurs during resize,
which seems to have to do with NSTextView wanting to center the
selection.)

When a Hemlock Frame (window) closes, set its autosave name to #@"".
(This allows previously used names - including listener document
names - to be reused.)

In GET-DEFAULT-ENCODING, only bother if CCL:*DEFAULT-CHARACTER-ENCODING*
is a keyword for which CCL::LOOKUP-CHARACTER-ENCODING returns non-nil.
(This increases the chance that GET-DEFAULT-ENCODING may return NIL.)

Don't expect an :INITFORM to take effect for a foreign slot in
the DEFCLASS for HEMLOCK-EDITOR-DOCUMENT; set the slot in the #/init
method instead.

In UPDATE-BUFFER-PACKAGE, fall back to the (new) default package if
there's no IN-PACKAGE form before point.

For now, the encodings popup menu in Open and Save dialogs only lists
the ~2 dozen encodings that CCL actually supports.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@12125 8bdb8f1d-59ce-db11-9e41-0016172a54ae