Add a HAVE-INTERACTIVE-TERMINAL-IO slot to APPKIT-PROCESS.  (This basically means that we're running in the hybrid (require "COCOA") environment, running the bundled executable directly in the shell, or have been able to connect to AltConsole.   Might later mean that we have a swank connection or similar.)
Experiment with other ways to do PROCESS-INTERRUPT on the AppKit thread;
code is not ready and may turn out to be a bad idea (conditionalized out.)

In EVENT-LOOP, wrap a restart (not a handler) around the #/run method call
if the event process has (some form of) interactive terminal i/o.

In the PROCESS-DEBUG-CONDITION method, enter a break loop if we have
interactive terminal i/o.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@12140 8bdb8f1d-59ce-db11-9e41-0016172a54ae