Revert DOLIST to the traditional version; hack declarations so that if there's a TYPE declaration on VAR, it's widened to (OR NULL type) when the iteration variable can be NIL and narrowed when it's known not to be.
This seems to ... um, further highlight ... our problems with
typechecking at SAFETY 3.  (Most of which are problems with SAFETY 3
being so ridiculous ...)

I think that more things that the spec says about DOLIST are likely
to be true in this expansion than in the one that we'd been using,
but wouldn't claim to really like this.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@12646 8bdb8f1d-59ce-db11-9e41-0016172a54ae