Change *SAVE-SOURCE-LOCATIONS* default from :NO-TEXT to T.  Leave *CCL-SAVE-SOURCE-LOCATIONS* as :NO-TEXT by default, but set it to T  in SET-DEVELOPMENT-ENVIRONMENT.
The reasoning is this: :no-text only makes sense for files that don't change.  For files that change, it's worse than useless.  For files that don't change, it works just fine and saves space.  For most users, the ccl sources are read-only, so :no-text is fine for building ccl.  Users who modify ccl sources should do (set-development-environment) in their init files to set up the best environment for doing so.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@12664 8bdb8f1d-59ce-db11-9e41-0016172a54ae