Add an "emulate" restart for a write-to-watched-object condition.
This operates at a very low level: it emulates the machine instruction
that caused the write.

In order to do this, we suspend other threads and suspend the gc so
that we can use the register state in the exception context to
compute an effective address and perform the store.

The code only knows how to emulate storing a GPR to memory; the
restart will not be offered if the faulting instruction is some other
kind of operation.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@13014 8bdb8f1d-59ce-db11-9e41-0016172a54ae