Revive some old code which tries to ensure that lisp slot-vectors are removed when NSObjects are deallocated.
This approach may have problems (because of the way that tls cleanup
works on OSX, or at least the way that it used to) and is suboptimal;
the comments near the change discuss some other approaches.

It'd be interesting to see whether this change causes the symptoms
describe in ticket:706 to disappear.  (If so, we know what the problem
is and may need to think of a better way to fix it.)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14198 8bdb8f1d-59ce-db11-9e41-0016172a54ae