l1-boot-2.lisp: In INITIALIZE-INTERACTIVE-STREAMS: ensure that *STDOUT*, *STDERR*, and  (if it's disjoint) *TERMINAL-OUTPUT* are auto-flushed.
linux-files.lisp:
#+windows-target Define GET-LAST-WINDOWS-ERROR, so that #_GetLastError
can be called from code that may have to run before the FFI is
initialized.
When prompting the user to type :Y, force output.  (This may run on the
housekeeping thread, so that output may not get autoflushed.)

l0-cfm-support.lisp: if GET-SHARED-LIBRARY fails, signal the error on
the calling thread (even if the #_dlopen/whatever happens on the initial
thread); fixes ticket:742


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14288 8bdb8f1d-59ce-db11-9e41-0016172a54ae