Pasting clipboard content from other applications doesn't work under Cocotron, because of confusion about the format and endianness of that data.  (What actually gets pasted is a string with byte-reversed character codes, which may not be visible even though it make take up space in the buffer.)
Try to work around that: if more than half of the (16-bit) characters
in the string obtained from the clipboard have 0s in their low bytes,
assume that the string is byte-reversed 8-bit text and reverse those
characters again.  This can be fooled, so this workaround should be
removed when the underlying Cocotron bug is fixed.

For the sake of argument, this fixes ticket:740, though we should
report the underlying Cocotron confusion and try to render the workaround
unnecessary.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14481 8bdb8f1d-59ce-db11-9e41-0016172a54ae