REMHASH: it's not our job to effectively rehash the table when the last key's removed.  (That should happen in grow/rehash code, anyway.)
%GROW-HASH-TABLE now calls new function %GROW-HASH-TABLE-IN-PLACE-P
to decide whether or not to grow the table or do an in-place rehash.
We might want to make this customizable; the default implementation
suggests an in-place rehash if there are at least half as many deleted
entries, and that seems to give good results for the test in ticket:809.

Fixes ticket:809, AFAICT.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14625 8bdb8f1d-59ce-db11-9e41-0016172a54ae