When computing the absolute value of a complex number, be a little more careful about it.
hypot in libm is even more careful, and one could
certainly make the argument that we should just arrange
to ff-call that.

This new version conses more than what we were
previously doing, but seems to be about the same speed.
If this ends up being a performance issue, writing a
LAP version probably would not be too tough.

Fixes ticket:830.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14673 8bdb8f1d-59ce-db11-9e41-0016172a54ae