When moving point to new position by mouse click, leave mark at the old position so c-x c-x (or c-u c-@) will bring you back.
Make c-x c-x with prefix argument deactivate region.

Make cmd-E invoke editor-execute-expression-command, in particular it doesn't require a selection.

Treat the colon in keywords as a word constituent for selection purposes.

Make extending the selection use the same rules as setting it.  This eliminates a number of inconsistencies in selection behavior.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14737 8bdb8f1d-59ce-db11-9e41-0016172a54ae