Pass -march=arm7-a and -mno-warn-deprecated to the assembler.
Without -march=armv7-a, the assembler warns about certain operand
restrictions with the mul and mla instructions.  These restrictions
were lifted as of armv6.  (Rd and Rn are supposed to be distinct
on earlier architecture versions.)

Unfortunately, with -march=armv7-a, the assembler gets all shirty
if you use the swp instruction, and warns.  We never use swp for
synchronization, and are not interested in this warning, so just
turn deprecation warnings off.



git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14782 8bdb8f1d-59ce-db11-9e41-0016172a54ae