Define and export the functions ALLOW-HEAP-ALLOCATION and HEAP-ALLOCATION-ALLOWED-P and the condition type ALLOCATION-DISABLED.
(ALLOW-HEAP-ALLOCATION arg) : when ARG is NIL, causes any subsequent
attempts to heap-allocate lisp memory to signal (as if by CERROR)
an ALLOCATION-DISABLED condition.  (Allocaton is enabled globally at
the point where the error is signaled.)  Continuing from the CERROR
restarts the allocation attempt.

This is intended to help verify that code that's not expected to
cons doesn't do so.

(This is only implemented on the ARM at the moment, but the intent
is that it be supported on all platforms.)

Note that calling (ALLOW-HEAP-ALLOCATION NIL) in the REPL CERRORs
immediately, since the REPL will cons to create the new value of CL:/.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14807 8bdb8f1d-59ce-db11-9e41-0016172a54ae