Save the (C) non-volatile FPRs (d8-d15, which are aliased to s16-s31) when creating CATCH/UNWIND-PROTECT frames; restore them appropriately when throwing/unwinding.  (Having lisp treat the C non-volatile FPRs as non-volatile means that we don't have to do anything special on callbacks.)
Most of these changes are transparent to compiled code, but try to
account for the presence of an extra stack-allocated object (a
DOUBLE-FLOAT vector) during unwind-protect cleanup.  That's all
intended to allow code like:

{{{
(tagbody
  (unwind-protect
       (some-protected-form)
    (go OUT))
  OUT
  ...)
}}}

to work, but I'm not sure that it ever has on the ARM.

Technically, this is an ABI change but it's obscure enough (and I'm
lazy enough) that I did't change FASL or image version numbers.
If you have code like the above, recompile it (or take it out back
and shoot it, if you prefer.)

Having the values of these FPRs perserved by CATCH and friends is
one prerequisite to using them in compiled lisp code.  (There's
a lot more to that, of course.)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14823 8bdb8f1d-59ce-db11-9e41-0016172a54ae